These properties should be long, to eliminate type mismatches

+++ hab/speedway/global.c
@@ -41,9 +41,9 @@ int set_antenna_epsilon = 0;
 int set_peakrssi_delta = 0;
 int set_peakrssi_epsilon = 0;
 
-unsigned int peakrssi_delta = 0;
-unsigned int peakrssi_epsilon = 0;
-unsigned int loss_timeout = UINT_MAX;
+unsigned long peakrssi_delta = 0;
+unsigned long peakrssi_epsilon = 0;
+unsigned long loss_timeout = ULONG_MAX;
 //int tag_direction = 0;
 
 LLRP_tSConnection *pConn;
